Moved structs and enums from the .c file to header
